fix: a scope reads every value applied to it, not the first seven
Where each applied value sits in the calldata was answered as a byte. The selector
comes first and each value takes a word after it, so the eighth sits at 256 — and
256 in a byte is zero. The eighth value read the selector, the ninth read the
first, and the contract answered a number rather than failing.

Measured before it was fixed: at eight arguments the chain answered a piece of the
selector where the evaluator answered 42, and at twelve it answered the fourth
argument's value. A wrong answer on a chain, which is the one kind of wrong this
backend may not be.

Two bytes, like the other PUSH1 ceilings before it, and it cost nothing anywhere
else: the prologue is measured by writing it rather than added up from a table, so
it adjusted itself.

Seven sizes in the harness, from one to twenty-four, and each asks twice — the
last value alone, which says where it was read from, and the sum of all of them,
which says every one was. Each argument is worth its own position, so a value read
from the wrong place is a wrong answer rather than the same answer by luck.

+++ builder/evm/getter.go
@@ -16,8 +16,13 @@ func GetRuntimeCodeLength(rc *RuntimeCode) int {
 	return l
 }
 
-// GetCalldataArgsOffset returns the calldata byte offset for the Nth argument (0-based).
-// ABI layout: selector at 0, then each arg in a 32-byte slot: arg0 at 0x20, arg1 at 0x40, arg2 at 0x60, ...
-func GetCalldataArgsOffset(index uint64) byte {
-	return byte(CALLDATA_SLOT_READABLE * (index + 1))
+// GetCalldataArgsOffset answers where the Nth value applied to a scope sits in the calldata,
+// counting from zero.
+//
+// The selector is first and each value has a word after it, so the eighth sits at 256 — which
+// is why this answers a number and not a byte. It used to answer a byte, and 256 in a byte is
+// zero: the eighth value read the selector, the ninth read the first, and the contract answered
+// something rather than failing.
+func GetCalldataArgsOffset(index uint64) int {
+	return CALLDATA_SLOT_READABLE * int(index+1)
 }
